backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
motor==3.1.1
pymongo==4.3.3
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==0.8.11
openai==1.12.0
httpx==0.26.0
python-jose==3.3.0
passlib==1.7.4 